+++ src/index.ts
@@ -15,3 +15,14 @@ export {wait, wrap} from './utils/async.js';
 export {last, toArray, EMPTY_ARRAY} from './utils/array.js';
 export {loadPackageJson} from './project/packageJson.js';
 export {TaskError} from './task/task.js';
+export type {Log, LoggerState} from './utils/log.js';
+export {
+	LogLevel,
+	Logger,
+	SystemLogger,
+	DevLogger,
+	ENV_LOG_LEVEL,
+	DEFAULT_LOG_LEVEL,
+	configureLogLevel,
+	printLogLabel,
+} from './utils/log.js';