+++ changelog.md
@@ -8,7 +8,7 @@
   for undefined Gro config properties:
   `GRO_HOST`, `GRO_PORT`, `GRO_LOG_LEVEL`
   ([#162](https://github.com/feltcoop/gro/pull/162))
-- document the dev logger and fix a bug in the two logger classes
+- export the logger API from root and document and fix it
   ([#162](https://github.com/feltcoop/gro/pull/162))
 
 ## 0.16.0